Исправления и дополнения UserService
